backend/requirements.txt
Flask
flask_cors
requests
beautifulsoup4
google-generativeai
python-dotenv